fastapi-auth/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
boto3==1.37.10
botocore==1.37.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
eventlet==0.39.1
falcon==4.0.2
fastapi==0.115.11
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
jmespath==1.0.1
MarkupSafe==3.0.2
mongoengine==0.29.1
packaging==24.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.1.3
